backend/requirements.txt
flask
flask-jwt-extended
flask-bcrypt
python-dotenv
flask-migrate
flask-sqlalchemy
flask-cors